Move ArrowWriter round trip tests to parquet/tests/arrow_writer/roundtrip.rs
Moves the round trip tests (write with ArrowWriter, read back, compare data)
from the tests module in parquet/src/arrow/arrow_writer/mod.rs to a new
integration test file parquet/tests/arrow_writer/roundtrip.rs.

The test_page_stats_* tests cannot move because they use crate-private
thrift APIs, so a copy of the roundtrip_opts helper is retained for them.

Part of https://github.com/apache/arrow-rs/issues/10540
